requirements.txt
paramiko
pyyaml
urllib2
